8278547: riscv: Refactor moving assembler functions for vector instructions into separate file
